refactor(ads): move types and clean up attributes
