+++ purchases/src/defaults/kotlin/com/revenuecat/purchases/ads/rewardverification/RewardVerificationToken.kt
@@ -1,4 +1,8 @@
-package com.revenuecat.purchases
+package com.revenuecat.purchases.ads.rewardverification
+
+import com.revenuecat.purchases.ExperimentalPreviewRevenueCatPurchasesAPI
+import com.revenuecat.purchases.Purchases
+import dev.drewhamilton.poko.Poko
 
 /**
  * Ties a loaded rewarded ad to its server-side reward verification.
@@ -8,11 +12,12 @@ package com.revenuecat.purchases
  * callback with [Purchases.pollRewardVerification].
  */
 @ExperimentalPreviewRevenueCatPurchasesAPI
-public data class RewardVerificationToken(
+@Poko
+public class RewardVerificationToken(
     /** Set as the ad network's server-side verification custom data. */
-    val customData: String,
+    public val customData: String,
     /** Correlates the ad with its verification. */
-    val clientTransactionId: String,
+    public val clientTransactionId: String,
     /** The app user the reward is attributed to; set as the ad network's SSV user identifier. */
-    val appUserID: String,
+    public val appUserID: String,
 )
